0.8.1
Move cli code into its own module.
