+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ink"
-version = "0.8.0"
+version = "0.8.1"
 edition = "2021"
 
 [profile.release]
